Патч #4 Сделано: * убрана заглушка и добавлена фича с фаворитами
